% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_area.R
\name{locate_areas}
\alias{extract_areas}
\alias{locate_areas}
\title{extract_areas}
\usage{
locate_areas(file, pages = NULL, silent = TRUE)

extract_areas(file, pages = NULL, guess = FALSE, ...)
}
\arguments{
\item{file}{A character string specifying the path to a PDF file. This can also be a URL, in which case the file will be downloaded to the R temporary directory using \code{download.file}.}

\item{pages}{An optional integer vector specifying pages to extract from. To extract multiple tables from a given page, repeat the page number (e.g., \code{c(1,2,2,3)}).}

\item{silent}{A logical indicating whether to silence the \code{\link[graphics]{locator}} function.}

\item{guess}{See \code{\link{extract_tables}} (note the different default value).}

\item{\dots}{Other arguments passed to \code{\link{extract_tables}}.}
}
\value{
For \code{extract_areas}, see \code{\link{extract_tables}}. For \code{locate_areas}, a list of four-element numeric vectors (top,left,bottom,right), one per page of the file.
}
\description{
Interactively identify areas and extract
}
\details{
\code{extract_areas} is an interactive mode for \code{\link{extract_tables}} allowing the user to specify areas of each PDF page in a file that they would like extracted. In interactive mode, each page is rendered to a PNG file and displayed in an R graphics window sequentially, pausing on each page to call \code{\link[graphics]{locator}} so the user can specify two points (e.g., upper-left and lower-right) to define bounds of page area. \code{extract_areas} then passes these user-defined areas to \code{\link{extract_tables}}. \code{locate_areas} implements the interactive component only, without actually extracting; this might be useful for interactive work that needs some modification before executing \code{extract_tables} computationally.
}
\note{
Currently, attempting to resize the graphics window at any point during this process will cause problems.
}
\examples{
\dontrun{
# simple demo file
f <- system.file("examples", "data.pdf", package = "tabulizer")

# locate areas only
locate_areas(f)

# locate areas and extract
extract_areas(f)
}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\seealso{
\code{\link{extract_tables}}, \code{\link{make_thumbnails}}, , \code{\link{get_page_dims}}
}

