% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metadata.R
\name{extract_metadata}
\alias{extract_metadata}
\title{extract_metadata}
\usage{
extract_metadata(file)
}
\arguments{
\item{file}{A character string specifying the path or URL to a PDF file.}
}
\value{
A list.
}
\description{
Extract metadata from a file
}
\details{
This function extracts metadata from a PDF
}
\examples{
\dontrun{
# simple demo file
f <- system.file("examples", "data.pdf", package = "tabulizer")

extract_metadata(f)
}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\seealso{
\code{\link{extract_tables}}, \code{\link{extract_areas}}, \code{\link{extract_text}}
}

