% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tables.R
\name{extract_tables}
\alias{extract_tables}
\title{extract_tables}
\usage{
extract_tables(file = "../inst/examples/mtcars.pdf", pages = NULL,
  area = NULL, columns = NULL, guess = TRUE, ...)
}
\arguments{
\item{file}{A character string specifying the path to a PDF file.}

\item{pages}{An optional integer vector specifying pages to extract from.}

\item{area}{An optional list, of length equal to the number of pages specified, where each entry contains a four-element numeric vector of coordinates (top,left,bottom,right) containing the table for the corresponding page. Specify \code{area} xor \code{columns}.}

\item{columns}{An optional list, of length equal to the number of pages specified, where each entry contains a numeric vector of horizontal (x) coordinates separating columns of data for the corresponding page. Specify \code{area} xor \code{columns}.}

\item{guess}{A logical indicating whether to guess the locations of tables on each page. If \code{FALSE}, \code{area} or \code{columns} must be specified.}

\item{\dots}{Ignored.}
}
\value{
A list of character matrices.
}
\description{
Extract tables from a file
}
\details{
This function mimics the behavior of the tabula command line utility. It returns a list of R data.frames containing tables extracted from a file.
}
\examples{
# simple demo file
f <- system.file("examples", "data.pdf", package = "tabulizer")

# extract all tables
extract_tables(f)

# extract tables from only second page
extract_tables(f, pages = 2)

}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
\href{http://tabula.technology/}{Tabula}
}

