% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_text.R
\name{extract_text}
\alias{extract_text}
\title{extract_text}
\usage{
extract_text(file, pages = NULL, password = NULL, encoding = NULL)
}
\arguments{
\item{file}{A character string specifying the path or URL to a PDF file.}

\item{pages}{An optional integer vector specifying pages to extract from.}

\item{password}{Optionally, a character string containing a user password to access a secured PDF.}

\item{encoding}{Optionally, a character string specifying an encoding for the text, to be passed to the assignment method of \code{\link[base]{Encoding}}.}
}
\value{
If \code{pages = NULL} (the default), a length 1 character vector, otherwise a vector of length \code{length(pages)}.
}
\description{
Extract text from a file
}
\details{
This function converts the contents of a PDF file into a single unstructured character string.
}
\examples{
\dontrun{
# simple demo file
f <- system.file("examples", "data.pdf", package = "tabulizer")

# extract all text from page 1 only
extract_text(f, from = 1, to = 1)

# extract all text
extract_text(f)
}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\seealso{
\code{\link{extract_tables}}, \code{\link{extract_areas}}, \code{\link{split_pdf}}
}

