% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_page_dims.R
\name{get_page_dims}
\alias{get_page_dims}
\title{get_page_dims}
\usage{
get_page_dims(file, pages = NULL)
}
\arguments{
\item{file}{A character string specifying the path or URL to a PDF file.}

\item{pages}{An optional integer vector specifying pages to extract from.}
}
\value{
A list of two-element numeric vectors specifying the width and height of each page, respectively.
}
\description{
Get Page Dimensions
}
\details{
This function extracts the dimensions of pages in a PDF document. This can be useful for figuring out how to specify the \code{area} argument in \code{\link{extract_tables}}
}
\examples{
\dontrun{
# simple demo file
f <- system.file("examples", "data.pdf", package = "tabulizer")

get_page_dims(f)
}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
\href{http://tabula.technology/}{Tabula}
}
\seealso{
\code{\link{extract_tables}}, \code{\link{extract_text}}, \code{\link{make_thumbnails}}
}

