% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_columns.R
\name{locate_columns}
\alias{locate_columns}
\title{Locate separators between columns}
\usage{
locate_columns(file, pages = NULL, resolution = 60L, copy = FALSE)
}
\arguments{
\item{file}{A character string specifying the path or URL to a PDF file.}

\item{pages}{An optional integer vector specifying pages to extract from.}

\item{resolution}{An integer specifying the resolution of the PNG images conversions. A low resolution is used by default to speed image loading.}

\item{copy}{Specifies whether the original local file(s) should be copied to tempdir() before processing. FALSE by default. The argument is ignored if file is URL.}
}
\value{
a list.
}
\description{
This function allows the user to manually locate the separators between columns of a table in a pdf. The output can be used as the \code{columns} argument in \code{extract_tables()}
}
\details{
Manually selecting the separators can ensure that values stay in their respective columns. This is useful when some rows of a table have no or only little white space between columns. The code is an adaptation of the \code{locate_area()} function and its helpers.
}
\examples{
\donttest{
f <- system.file("examples", "data.pdf", package = "tabulizer")
separators<-locate_columns(f, pages= 1 )
extract_tables(f,pages = 1, columns = separators[1])
}

}
\author{
Thore Engel <thore.engel@posteo.de>
}
