% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_thumbnails.R
\name{make_thumbnails}
\alias{make_thumbnails}
\title{make_thumbnails}
\usage{
make_thumbnails(file, pages = NULL, format = c("png", "jpeg", "bmp", "gif"),
  resolution = 72L)
}
\arguments{
\item{file}{A character string specifying the path to a PDF file.}

\item{pages}{An optional integer vector specifying pages to extract from.}

\item{format}{A character string specifying an image file format.}

\item{resolution}{An integer specifying the image resolution in DPI.}
}
\value{
A character vector of file paths.
}
\description{
Convert Pages to Image Thumbnails
}
\details{
This function save each (specified) page of a document as an image with 720 dpi resolution. Images are saved in the same directory as the original file, with file names specified by the original file name, a page number, and the corresponding file format extension.
}
\note{
This may generate Java \dQuote{INFO} messages in the console, which can be safely ignored.
}
\examples{
\dontrun{
# simple demo file
f <- system.file("examples", "data.pdf", package = "tabulizer")

make_thumbnails(f)
}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
\href{http://tabula.technology/}{Tabula}
}
\seealso{
\code{\link{extract_tables}}, \code{\link{make_thumbnails}}
}

