% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_merge.R
\name{split_pdf}
\alias{merge_pdfs}
\alias{split_pdf}
\title{Split and merge PDFs}
\usage{
split_pdf(file)

merge_pdfs(file, outfile)
}
\arguments{
\item{file}{For \code{merge_pdfs}, a character vector specifying the path to one or more local PDF files. For \code{split_pdf}, a character string specifying the path or URL to a PDF file.}

\item{outfile}{A character string specifying the path to the PDF file to create from the merged documents.}
}
\value{
For \code{split_pdfs}, a character vector specifying the output file names, which are patterned after the value of \code{file}. For \code{merge_pdfs}, the value of \code{outfile}.
}
\description{
Split PDF into separate pages or merge multiple PDFs into one.
}
\details{
\code{\link{split_pdf}} splits the file listed in \code{file} into separate one-page doucments. \code{\link{merge_pdfs}} creates a single PDF document from multiple separate PDF files.
}
\examples{
\dontrun{
# simple demo file
f <- system.file("examples", "data.pdf", package = "tabulizer")
get_n_pages(file = f)

# split PDF by page
sf <- split_pdf(f)

# merge pdf
merge_pdfs(sf, "merged.pdf")
get_n_pages(file = "merged.pdf")
}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\seealso{
\code{\link{extract_areas}}, \code{\link{get_page_dims}}, \code{\link{make_thumbnails}}
}

