% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_columns.R
\name{try_columns_reduced}
\alias{try_columns_reduced}
\title{Helper function to locate_columns()}
\usage{
try_columns_reduced(file, dims, warn = FALSE)
}
\arguments{
\item{file}{}

\item{dims}{}

\item{warn}{}
}
\description{
Helper function to locate_columns()
}
