% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{stop_logging}
\alias{stop_logging}
\title{rJava logging}
\usage{
stop_logging()
}
\value{
\code{NULL}, invisibly.
}
\description{
Toggle verbose rJava logging
}
\details{
This function turns off the default, somewhat verbose rJava logging, most of which is uninformative.
}
\note{
This resets a global Java setting and may affect logging of other rJava operations, requiring a restart of R.
}
\examples{
\dontrun{
stop_logging()
}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}

