% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_columns.R
\name{try_columns_reduced}
\alias{try_columns_reduced}
\title{Helper function to locate_columns()}
\usage{
try_columns_reduced(file, dims, warn = FALSE)
}
\arguments{
\item{file}{A character string specifying the path or URL to a PDF file.}

\item{dims}{An integer specifying the resolution of the PNG images conversions. A low resolution is used by default to speed image loading.}

\item{warn}{Display warning?}
}
\description{
Helper function to locate_columns()
}
